Add link to the loyalty API doc
Add a link to the API developer's guide section about building a loyalty integration. We have an [open request to make this reference implementation public](https://toasttab.atlassian.net/browse/OPX-2669) so that external developers can use it. When the repository is public, external readers will need a link to some information about what the implementation is for.

@toasttab/core-integrations, @toasttab/open-source, would you take a look at this minor addition to the README.md file?